requirements.txt
django
djangorestframework
django-jazzmin